tidy interpret-expression modules
