+++ packages/compiler/src/parser/interpret-expression/token-state.ts
@@ -1,0 +1,30 @@
+import { ExpressionToken } from '../produce-expression-tokens';
+
+export interface WithTokens<T> {
+  tokens: ExpressionToken[];
+  value: T;
+}
+
+export function withTokens<T>(tokens: ExpressionToken[], value: T): WithTokens<T> {
+  return { tokens, value };
+}
+
+export function mapWithTokens<A, B>(f: (a: A) => B): (result: WithTokens<A>) => WithTokens<B>;
+export function mapWithTokens<A, B>(result: WithTokens<A>, f: (a: A) => B): WithTokens<B>;
+export function mapWithTokens<A, B>(result: WithTokens<A> | ((a: A) => B), f?: ((a: A) => B)): WithTokens<B> | ((result: WithTokens<A>) => WithTokens<B>) {
+  if (typeof result === 'function') {
+    return (actualResult) => mapWithTokens(actualResult, result);
+  }
+
+  if (!f) {
+    throw new Error('Missing parameter to mapWithTokens');
+  }
+
+  const { tokens, value } = result;
+  return withTokens(tokens, f(value));
+}
+
+export function flatMapWithTokens<A, B>({ tokens, value }: WithTokens<A>, f: (a: A) => WithTokens<B>): WithTokens<B> {
+  const { tokens: resultTokens, value: resultValue } = f(value);
+  return withTokens([...tokens, ...resultTokens], resultValue);
+}
