+++ packages/compiler/src/parser/parse.ts
@@ -1,5 +1,6 @@
 import { Expression } from '..';
-import interpretExpression, { WithMessages } from './interpret-expression';
+import interpretExpression from './interpret-expression/interpreters';
+import { WithMessages } from './interpret-expression/message-state';
 import { produceExpressionTokens } from './produce-expression-tokens';
 import tokenize from './tokenize';
